Move person search component and context
