+++ frontend/src/employee-frontend/App.tsx
@@ -30,7 +30,6 @@ import MobilePairingModal from './components/MobilePairingModal'
 import PersonProfile from './components/PersonProfile'
 import PersonalMobileDevicesPage from './components/PersonalMobileDevicesPage'
 import Reports from './components/Reports'
-import Search from './components/Search'
 import UnitFeaturesPage from './components/UnitFeaturesPage'
 import UnitPage from './components/UnitPage'
 import Units from './components/Units'
@@ -67,6 +66,7 @@ import InvoicesPage from './components/invoices/InvoicesPage'
 import MessagesPage from './components/messages/MessagesPage'
 import OutOfOfficePage from './components/out-of-office/OutOfOfficePage'
 import PaymentsPage from './components/payments/PaymentsPage'
+import Search from './components/person-search/Search'
 import PlacementDraftPage from './components/placement-draft/PlacementDraft'
 import PlacementToolPage from './components/placement-tool/PlacementToolPage'
 import ReportApplications from './components/reports/Applications'
